Fix SyncBatchNorm for empty inputs
fixes #36530

Prior to this commit, SyncBatchNorm crashes with the following
error message.

```
File "..../torch/nn/modules/_functions.py", line 17, in forward
    mean, invstd = torch.batch_norm_stats(input, eps)
RuntimeError: cannot reshape tensor of 0 elements into shape [0, 3, -1] because the unspecified dimension size -1 can be any value and is ambiguous
```

This PR adds a dedicated branch to handle empty inputs. When a process
recieves empty inputs, it will set its local `mean`, `invstd`, and `count`
to zero, and participate in the `all_gather` collective communications in
the forward pass. Then `mean` and `invstd` with zero count will be
filtered out before computing global mean and invstd. In the backward
pass, it also participate in the `all_reduce` communication with zero
tensors to unblock its peers.

ghstack-source-id: b060e5122937631e30d850c14690adc4b9251700
Pull Request resolved: https://github.com/pytorch/pytorch/pull/74944